добавлены кнопки-ссылки перехода к вопросам по статусу в таблице категорий, почистил код

+++ resources/views/categories/questions.blade.php
@@ -1,53 +1,64 @@
 @extends('dashboard')
 @section('item')
-<h2>{{ $category->name }}</h2>
+
+<h2 class="h2">{{ $category->name }}</h2>
 
 <table class="table table-hover">
 	<thead>
 	  <tr>
-	    <th></th>
-		<th>Автор</th>
-		<th>Email</th>
-		<th>Дата создания</th>
-		<th>Статус</th>
-		<th></th>
+	    <th class="w-10"></th>
+		<th class="text-center w-15">Автор</th>
+		<th class="text-center w-15">Email</th>
+		<th class="text-center w-15">Дата создания</th>
+		<th class="text-center w-12">Статус</th>
+		<th class="w-33"></th>
 	  </tr>
 	</thead>
 	<tbody>
-		@foreach ($category->questions as $question)
+		@foreach ($questions as $question)
 			<tr class="table-primary">
 			    <td>Инфо</td>
-			   	<td>{{ $question->author }}</td>
-				<td>{{ $question->email }}</td>
-				<td>{{ $question->created_at }}</td>
-				<td>{{ $status[$question->status] }}</td>
-				<td></td>
+			   	<td class="text-center">{{ $question->author }}</td>
+				<td class="text-center">{{ $question->email }}</td>
+				<td class="text-center">{{ $question->created_at }}</td>
+				<td class="text-center">{{ $status[$question->status] }}</td>
+				<td class="text-center"></td>
 			</tr>
 			<tr class="table-success">
-			    <td>Вопрос</td>
+			    <td>Вопрос:</td>
 			    <td colspan="4">{{ $question->name }}</td>
-			    <td><a href="/question/{{ $question->id }}/edit" class="btn btn-info mx-2 float-left">Изменить</a>
-					{!! Form::open(['url' => '/question/'.$question->id, 'method' => 'delete']) !!}
-					{!! Form::submit('Удалить', ['class' => 'btn btn-danger mx-2']) !!}
-					{!! Form::close() !!}
+			    <td>
+			        <div class="row">
+			            
+			            @if (empty($question['answer']))
+        					    <a href="/question/{{ $question->id }}" class="btn btn-info mx-1">Ответить</a>
+        				@endif
+			            
+    			        <a href="/question/{{ $question->id }}/edit" class="btn btn-info mx-1">Изменить</a>
+    				
+    					{!! Form::open(['url' => '/question/'.$question->id, 'method' => 'delete']) !!}
+    					{!! Form::submit('Удалить', ['class' => 'btn btn-danger mx-1']) !!}
+    					{!! Form::close() !!}
+    					
+    				</div>
 				</td>
 			</tr>
-			    
-    			    <tr>
-    			        <td>Ответ</td>
-    				    <td colspan="4">{{ $question->answer['answer'] }}</td>
-    				    <td>
-    				    	<a href="/answer/{{ $question->answer['id'] }}/edit" class="btn btn-info mx-2 float-left">Изменить</a>
-        					
-	    					{!! Form::open(['url' => '/answer/'.$question->answer['id'], 'method' => 'delete']) !!}
-	    					
-	    					{!! Form::submit('Удалить', ['class' => 'btn btn-danger mx-2']) !!}
-
-	    					{!! Form::close() !!}
-
-        				</td>
-    			    <tr>
-		        		
+			@if (!empty($question['answer']))
+    		    <tr>
+    		        <td>Ответ</td>
+    			    <td colspan="4">{{ $question->answer['answer'] }}</td>
+    			    <td>
+    			        <div class="row">
+    				    	<a href="/answer/{{ $question->answer['id'] }}/edit" class="btn btn-info mx-1">Изменить</a>
+        		
+        					{!! Form::open(['url' => '/answer/'.$question->answer['id'], 'method' => 'delete']) !!}
+        					{!! Form::hidden('url', $_SERVER['REQUEST_URI']) !!}
+        					{!! Form::submit('Удалить', ['class' => 'btn btn-danger mx-1']) !!}
+        					{!! Form::close() !!}
+                        </div>
+    				</td>
+    		    <tr>
+    		@endif
 		@endforeach
 	</tbody>
 </table>
